+++ src/biz/bokhorst/xprivacy/ActivityMain.java
@@ -32,6 +32,7 @@ import android.os.Build;
 import android.os.Bundle;
 import android.os.Environment;
 import android.os.Handler;
+import android.support.v4.content.LocalBroadcastManager;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.util.Log;
@@ -67,6 +68,10 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 	private boolean mFiltersHidden = true;
 	private boolean mCategoriesHidden = true;
 	private Bitmap[] mCheck;
+	private int mProgressWidth;
+	private int mProgress = 0;
+	private boolean mSharing = false;
+	private String mSharingState = null;
 
 	private static final int ACTIVITY_LICENSE = 0;
 	private static final int ACTIVITY_EXPORT = 1;
@@ -263,6 +268,13 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 			}
 		});
 
+		// Handle post share operation done message
+		if (mSharingState != null) {
+			TextView tvState = (TextView) findViewById(R.id.tvState);
+			tvState.setText(mSharingState);
+			mSharingState = null;
+		}
+
 		// Start task to get app list
 		AppListTask appListTask = new AppListTask();
 		appListTask.executeOnExecutor(mExecutor, (Object) null);
@@ -308,6 +320,7 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 		super.onDestroy();
 		if (mPackageChangeReceiver != null)
 			unregisterReceiver(mPackageChangeReceiver);
+		LocalBroadcastManager.getInstance(this).unregisterReceiver(progressListener);
 	}
 
 	@Override
@@ -354,16 +367,22 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 				}
 			}
 		} else if (requestCode == ACTIVITY_EXPORT) {
-			// Exported: send share intent
+			// Exported: clean-up UI
+			sharingDone();
+
+			// send share intent
 			if (data != null && data.hasExtra(ActivityShare.cFileName)) {
 				Intent intent = new Intent(android.content.Intent.ACTION_SEND);
 				intent.setType("text/xml");
 				intent.putExtra(Intent.EXTRA_STREAM,
 						Uri.parse("file://" + data.getStringExtra(ActivityShare.cFileName)));
-				startActivity(Intent.createChooser(intent, getString(R.string.app_name)));
+				startActivity(Intent.createChooser(intent,
+						String.format(getString(R.string.msg_saved_to), data.getStringExtra(ActivityShare.cFileName))));
 			}
 		} else if (requestCode == ACTIVITY_IMPORT) {
-			// Imported: recreate UI
+			// Imported: clean-up UI
+			sharingDone();
+			// recreate UI
 			ActivityMain.this.recreate();
 		} else if (requestCode == ACTIVITY_IMPORT_SELECT) {
 			// Result for import file choice
@@ -377,7 +396,9 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 					Util.bug(null, ex);
 				}
 		} else if (requestCode == ACTIVITY_FETCH) {
-			// Fetched: recreate UI
+			// Fetched: clean-up UI
+			sharingDone();
+			// recreate UI
 			ActivityMain.this.recreate();
 		}
 	}
@@ -394,8 +415,9 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 		boolean pro = (Util.hasProLicense(this) != null);
 		boolean mounted = Environment.MEDIA_MOUNTED.equals(Environment.getExternalStorageState());
 
-		menu.findItem(R.id.menu_export).setEnabled(pro && mounted);
-		menu.findItem(R.id.menu_import).setEnabled(pro && mounted);
+		menu.findItem(R.id.menu_export).setEnabled(pro && mounted && !mSharing);
+		menu.findItem(R.id.menu_import).setEnabled(pro && mounted && !mSharing);
+		menu.findItem(R.id.menu_fetch).setEnabled(!mSharing);
 		menu.findItem(R.id.menu_pro).setVisible(!pro);
 
 		return super.onPrepareOptionsMenu(menu);
@@ -520,12 +542,18 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 			CheckBox cbFSystem = (CheckBox) findViewById(R.id.cbFSystem);
 			ProgressBar pbFilter = (ProgressBar) findViewById(R.id.pbFilter);
 			TextView tvStats = (TextView) findViewById(R.id.tvStats);
+			TextView tvState = (TextView) findViewById(R.id.tvState);
 			String filter = String.format("%s\n%b\n%b\n%b\n%b\n%b\n%b\n%b", etFilter.getText().toString(),
 					cFbUsed.isChecked(), cbFInternet.isChecked(), cbFRestriction.isChecked(),
 					cbFRestrictionNot.isChecked(), cbFPermission.isChecked(), cbFUser.isChecked(),
 					cbFSystem.isChecked());
 			pbFilter.setVisibility(ProgressBar.VISIBLE);
 			tvStats.setVisibility(TextView.GONE);
+
+			// Adjust progress state width
+			RelativeLayout.LayoutParams tvStateLayout = (RelativeLayout.LayoutParams) tvState.getLayoutParams();
+			tvStateLayout.addRule(RelativeLayout.LEFT_OF, R.id.pbFilter);
+
 			mAppAdapter.getFilter().filter(filter);
 		}
 	}
@@ -627,6 +655,8 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 	}
 
 	private void optionExport() {
+		sharingStart();
+
 		Intent file = new Intent(Intent.ACTION_GET_CONTENT);
 		file.setType("file/*");
 		boolean multiple = Util.isIntentAvailable(ActivityMain.this, file);
@@ -636,6 +666,8 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 	}
 
 	private void optionImport() {
+		sharingStart();
+
 		Intent file = new Intent(Intent.ACTION_GET_CONTENT);
 		file.setType("file/*");
 		if (Util.isIntentAvailable(ActivityMain.this, file)) {
@@ -652,6 +684,8 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 	}
 
 	private void optionFetch() {
+		sharingStart();
+
 		if (Util.getProLicense() == null) {
 			// Redirect to pro page
 			Intent browserIntent = new Intent(Intent.ACTION_VIEW, cProUri);
@@ -988,10 +1022,16 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 			protected void publishResults(CharSequence constraint, FilterResults results) {
 				clear();
 				TextView tvStats = (TextView) findViewById(R.id.tvStats);
+				TextView tvState = (TextView) findViewById(R.id.tvState);
 				ProgressBar pbFilter = (ProgressBar) findViewById(R.id.pbFilter);
 				pbFilter.setVisibility(ProgressBar.GONE);
 				tvStats.setVisibility(TextView.VISIBLE);
 				tvStats.setText(results.count + "/" + AppListAdapter.this.mListApp.size());
+
+				// Adjust progress state width
+				RelativeLayout.LayoutParams tvStateLayout = (RelativeLayout.LayoutParams) tvState.getLayoutParams();
+				tvStateLayout.addRule(RelativeLayout.LEFT_OF, R.id.tvStats);
+
 				if (results.values == null)
 					notifyDataSetInvalidated();
 				else {
@@ -1256,6 +1296,58 @@ public class ActivityMain extends Activity implements OnItemSelectedListener, Co
 
 			return convertView;
 		}
+	}
+
+	// Share operations progress listener
+
+	private void sharingStart() {
+		IntentFilter filter = new IntentFilter();
+		filter.addAction(ActivityShare.cProgressReport);
+		LocalBroadcastManager.getInstance(this).registerReceiver(progressListener, filter);
+
+		mSharing = true;
+		invalidateOptionsMenu();
+
+		// Set up the progress bar
+		final View vProgressEmpty = (View) findViewById(R.id.vProgressEmpty);
+		mProgressWidth = vProgressEmpty.getMeasuredWidth();
+	}
+
+	private void sharingDone() {
+		// Re-enable menu items
+		mSharing = false;
+		invalidateOptionsMenu();
+		// Stop listening for progress reports
+		LocalBroadcastManager.getInstance(this).unregisterReceiver(progressListener);
+		// Keep done message for after UI recreation
+		TextView tvState = (TextView) findViewById(R.id.tvState);
+		mSharingState = (String) tvState.getText();
+	}
+
+	private BroadcastReceiver progressListener = new BroadcastReceiver() {
+		@Override
+		public void onReceive(Context context, Intent intent) {
+			// String action = intent.getAction();
+			String message = intent.getStringExtra(ActivityShare.cProgressMessage);
+			int progress = intent.getIntExtra(ActivityShare.cProgressValue, 0);
+			int max = intent.getIntExtra(ActivityShare.cProgressMax, 1);
+			setProgress(message, progress, max);
+		}
+	};
+
+	private void setProgress(String text, int progress, int max) {
+		TextView tvState = (TextView) findViewById(R.id.tvState);
+		if (text != null)
+			tvState.setText(text);
+		if (max == 0)
+			max = 1;
+		mProgress = (int) ((float) mProgressWidth) * progress / max;
+		updateProgress();
+	}
+
+	private void updateProgress() {
+		View vProgressFull = (View) findViewById(R.id.vProgressFull);
+		vProgressFull.getLayoutParams().width = mProgress;
 	}
 
 	// Helper methods