+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Changelog
 * Fixed Android KitKat message permissions
 * Do not disable Android/extra usage data when not in expert mode
 * Separate setting to enable restricting dangerous functions, thanks @[jpeg729](https://github.com/jpeg729)
+* Show exported file name while sharing, thanks @[jpeg729](https://github.com/jpeg729)
+* Progress indication for export, import and fetch, thanks @[jpeg729](https://github.com/jpeg729)
 * Added tri-state check box to legend ([issue](https://github.com/M66B/XPrivacy/issues/827))
 * Updated Lithuanian translation
 * Updated Norwegian translation